load_to_bq/requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
google-cloud-bigquery
